+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = kksubs
-version = 2.2
+version = 2.3
 
 [options]
 include_package_data = True
